Remove --with-metadata-cache and all code related to it.
This feature was improved to make rmlint scale over very large file collections.
However, after introducing the pathtricia trie, the memory advantage was only
a few percent while the cpu usage increased extremely sometimes.

Apart from that, the code was mangled all over the place and was hard to test
and add features that are compatible with --with-metadata-cache.