Prevent app window from getting stuck in fullscreen mode.
After switching to fullscreen mode, the app window loses its title bar,
making it impossible to get out of this state. This is the case with
Xfce, but it may also apply to other desktop environments. Maximising
the window instead of going fullscreen ensures we never get into this
state.