sh: Fix mistake in the `sh' formatting of the progress output
The format was not correct; the original intent was clearly to
pad the percentage field with space characters, which means that
no padding character should be specified, as the space character
is the default.

In addition, printf is now given a static formatting string, which
is at least safer in theory.