lib/{session.h,config.h.in}: Force inlining of abort functions.
Under modern semantics and historical practice, the function specifier
`inline' basically means "This function may be defined in multiple
translation units"; the idea is to give the compiler access to the
definition so that it can make better decisions about optimization,
particularly by inserting the body of the function into the call
site directly.

This means that the compiler is pretty much free to ignore `inline',
or even to pretend that a function not marked `inline' is worthy of
similar optimization. For the most part, this is a good thing; with
regard to optimization for the target machine, a compiler tends to
make much better decisions than a human.

Nevertheless, sometimes a human knows best, or at least a human
may desire to ensure the outcome one way or the other; sometimes,
the heuristics that a compiler use to make its decisions are set
up such that a small, seemingly insignificant change to the source
code might trigger the compiler to forego inlining. Thus, there
are 2 desirable features for the build process:

* The ability to receive warnings when a function that is declared
  `inline' is not in fact inlined; this commit turns on such warnings
  by using the following compiler option:

      -Winline

* The ability to force the compiler to inline a function even against
  its heuristical recommendations. A sophisticated compiler, such as
  GCC, Clang, or even MSVC has a language extension for this purpose;
  this commit employs this extension by defining the following macro:

      INLINE

The idea is that if a warning is ever produced for an inline function
that is not inlined, then that function or its usage can be investig-
ated, thereby allowing the programmer to decide whether it should be
re-written, or be declared without `inline', or indeed be forced via
`INLINE'.

As of this commit, no warning is produced for any inline function;
the macro `INLINE' has mainly been introduced here as a foundation
for further as-yet-complete work which does seem to require it.
However, for the sake of creating an air of immediate practicality,
`INLINE' is now used anyway when defining the following functions in
`lib/session.h', because they should always be inlined:

    rm_session_abort()
    rm_session_was_aborted()