po/SConscript: Run `xgettext' with `--join-existing' and `--add-location=file'
The `--join-existing'  option makes  `xgettext' try harder  to integrate
information  into the  existing  `rmlint.pot' file,  thereby easing  the
interpretation  of  the  diff;  otherwise,  `xgettext'  has  no  problem
re-arranging everything, and thereby making it impractical to figure out
what has changed.

The `--add-location=file' option prevents `xgettext' from outputing line
numbers in the comments that describe the  location of a string in the C
source code; while it may be useful to know this information, it changes
every time there is a shift  in the line location, and `--join-existing'
also preserves  old information, which  is just annoying. With  this new
option, only  the file name  is recorded, which  should cut down  on the
churn and still provide an indication that is good enough.