tree merge: Simplify memory allocation, and handle errors.
There's no reason to zero out `path_vec', as all the memory
but the last element is supposed to be overwritten in the
loop. Also, rather than abort on failure to allocate memory
(which is what happens when using `g_malloc0()'), the error
is now propagated up the stack and out to the user as a
failure status.