lib/treemerge.c: Store strings in the right order.
The  paths in  the  singly-linked list  `cfg->paths'  are stored  in
reverse order (they are prepended  onto the list essentially as they
are  encountered on  the  command line).  Though  it doesn't  really
matter,  `rm_tm_count_files()'  now  processes those  paths  in  the
correct order, simply by filling the `path_vec' array in reverse; at
the  very  least, this  has  the  benefit  of allowing  the  overall
function slightly  more efficient in terms  of variable assignements
and so on, as evidenced  by the commenting-out of the initialization
clause of the last `for' loop.