Use explicit types for some integer constants
The following constant expression:

    1024 * 1024

should be evaluated by the compiler at runtime as though it were:

    (int)1024 * (int)1024

and one would expect the result to be:

    (int)1048576

However, strictly speaking, an `int'  is only required to be able
to represent  at least  the number 32767.  However, a  `long int'
must be able  to represent at least  `2147483647'. Thus, strictly
speaking, proper C  code should specify that at least  one of the
integer constants is to be represented as a `long int':

    1024L * 1024

Consider, also, the following program:

    #include <limits.h> // UINT_MAX
    #include <stdio.h>  // printf

    #define PRINT_SIZE_OF(type) \
        printf("sizeof(" #type ") = %u\n", (unsigned int)sizeof(type))

    #define PRINT_UNSIGNED_LONG_LONG_INT(expr) \
        printf(#expr " = %llu\n", (unsigned long long int)(expr))

    int main()
    {
        PRINT_SIZE_OF(int);
        PRINT_SIZE_OF(long long int);

        PRINT_UNSIGNED_LONG_LONG_INT(UINT_MAX + 1);
        PRINT_UNSIGNED_LONG_LONG_INT(UINT_MAX + 1ULL);

        PRINT_UNSIGNED_LONG_LONG_INT(65536U * 65536);
        PRINT_UNSIGNED_LONG_LONG_INT(65536ULL * 65536);
    }

That program might produce the following output:

    sizeof(int) = 4
    sizeof(long long int) = 8
    UINT_MAX + 1 = 0
    UINT_MAX + 1ULL = 4294967296
    65536U * 65536 = 0
    65536ULL * 65536 = 4294967296

As you can see, specifying the type of an integer constant can be
important. Take, for example, one of the above calculations:

    65536U * 65536

Without the "U"  suffix to specify that the  integer constant has
an unsigned type, the compiler would compute:

    (int)65536 * (int)65536

On the  same machine, this would  result in overflow of  a signed
integer type,  which invokes undefined behavior;  a good compiler
would probably warn about it, or even produce an error.