lib/cmdline.c: Move constant loop condition (`paths') out of the loop
The compiler  probably already does  this, but I find  it clearer
and more  satisfying to do  this explicitly, especially  since it
allows avoiding a call to  `g_strfreev()', rather than relying on
that function to check `paths' for us again.

For the  sake of clarity in  the diff, the existing  code has not
yet been indented to reflect the  fact that it's part of the `if'
statement's consequent.