Assert `cfg->path_count == g_slist_length(cfg->paths)'
The variable `path_count'  is a bit of an  accounting hack rather
than  a  clean,  dependable  concept  (which  is  fine;  it's  an
implementation  detail). As  a result,  there is  the possibility
that it will be used incorrectly, as has already occurred:

  $ git checkout --detach 017b007e27374686d60272f18aa92064babd4480 &&
    git clean -fdx && scons DEBUG=1 && touch /tmp/{a.json,b} &&
    ./rmlint --replay /tmp/a.json --dedupe /tmp/b
  [...]
  ERROR:lib/session.c:189:rm_session_dedupe_main: assertion failed: (cfg->paths->next)
  ERROR: Aborting due to a fatal error. (signal received: Aborted)
  ERROR: Please file a bug report (See rmlint -h)

This programming mistake is the result of using `cfg->path_count'
as  though it  represents `g_slist_length(cfg->paths)',  which is
not the case when `--replay' is in effect. After this commit, the
mistake is clearer:

  $ ./rmlint --replay /tmp/a.json --dedupe /tmp/b
  ERROR:lib/session.c:174:rm_session_dedupe_main: assertion failed: (cfg->path_count == g_slist_length(cfg->paths))
  ERROR: Aborting due to a fatal error. (signal received: Aborted)
  ERROR: Please file a bug report (See rmlint -h)

The  new assertion  not only  documents an  expectation but  also
serves as a regression test.