xxhash: Use `g_slice_alloc()' and `g_slice_free1()'
The hash source  code provides a means  by which to plug  in your own
memory management routines, instead of using `malloc()' and `free()':

    XXH_malloc()
    XXH_free()

Implementing  `XXH_malloc()' is as simple  using `g_slice_alloc()' in
place of  `malloc()'. However,  `g_slice_free1()' must be  passed the
size of the object whose memory is being freed; as a result, in order
to implement  `XXH_free()' without resorting to  modifications of the
call site, it has to be implemented as a function-like macro, so that
the `sizeof'  operator can  be called  at the call  site in  order to
determine the requisite size. The only  reason this works is that the
call site uses a pointer to the known type of the object in question,
and thus the `sizeof' operator works properly, but it should be noted
that if the  call site is ever  modified so as not to  use the proper
type of the object, then this code will surely break.