Travis CI: Build with `DEBUG=1 O=release'
With the `DEBUG=1' option set,  the build entails extra tests for
bugs; in particular, the  build entails run-time assertions, such
as calls to `g_assert()'.

By default,  a debug  build undergoes no  optimization, so  as to
simplify and  speed up the design/build/test  cycle. However, for
the purposes of  CI, it may be  valuable to test a  build that is
optimized  similarly  to  a  release  build;  hence,  the  option
`O=release' is also used.