scons: Use `CCFLAGS' instead of `CFLAGS'
The rest of `SConstruct' uses  `CCFLAGS', so the one occurence of
`CFLAGS' has been changed to match.

The `CCFLAGS'  variable is  used for  both C  and C++,  while the
`CFLAGS' variable is  used for only C; at this  point in time, it
doesn't really  matter which  one is used,  but I've  come across
aspirations to port this program to Microsoft Windows, a platform
that effectively  provides no  "native" support  for C  (but does
provide "native" support for C++).