xxhash: `s/malloc/new/' and `s/free/delete/'
These identifiers:

    XXH_malloc
    XXH_free

are now:

    XXH_new
    XXH_delete

This helps  distinguish their semantics,  which may be as  flexible as
those similarly named  operators (`new' and `delete')  in C++; namely,
`delete' can handle a size argument in addition to a pointer argument.

In addition,  `XXH_malloc' has been  turned into  a macro so  that its
argument can be a type.