sh: {hard,sym}link: move dupe to temp file before attempting linking.
This acts as safety net when the target filesystem (where the dupe
lives) does not support symlinking (or hardlinking). Then we move the
dupe back to where it was instead of screwing things up.

See also Issue #429.

Also fixes a display bug that results in bad progress reports with -D.