cfg.c: fix bug introduced in 2.10.0
This reversion back to the code from version `2.9.0` addresses #438. Obviously there is some functionality here that's being implemented that I'm not familiar enough with the project to understand, but hopefully this gives someone a head start in trying to track down the bug, while still keeping the intended new behavior.