Clarify description and results of --is-reflink
Mention that a symlink passed to --is-reflink is reported as an error,
and that realpath(1) is a good way to preprocess arguments if either may
be a symlink to a candidate file.

"Are hardlinked/reflinked" sounds more like it identifies a relationship
between the files than "are hardlinks/reflinks", to me at least.

"Symlink" as a link type could imply that a symlink relationship is
identified. But "symlink" is an error status, so make that clear with
the word "encountered".

"Not a file" is inaccurate - use the term "regular file".