utils: rm_offset_get_from_fd: Don't combine separate extents
rmlint previously ignored holes between physically adjacent extents.
This could cause files to be detected as reflinked that weren't, if one
file had a complete extent and the other file had two pieces of it
separated by a hole.

Check for logical gaps as well as physical gaps to fix this problem.

Fixes #530