sh: relax skip_hardlink check
Apparently when --merge-directories is enabled, and --rank-by is not
sufficient to pick a specific original file in a group, the head
hardlink and the original file may not be the same. Relax the check in
rm_sh_emit_handler_hardlink to just compare the inode (the device is
already known to be the same).

Fixes #545