gui: actually make Polkit optional
If no versions of Polkit are available, skip the `gi.require_version`
call, or it will fail with ValueError even though it's an optional
dependency. This is necessary for the GUI to work on Homebrew and
Cygwin, which do not provide Polkit. Polkit isn't currently being used
for anything anyway.

Fixes this error:

Traceback (most recent call last):
  File "/usr/local/lib/python3.9/site-packages/shredder/__init__.py", line 11, in <module>
    gi.require_version('Polkit', '1.0')
  File "/usr/lib/python3.9/site-packages/gi/__init__.py", line 129, in require_version
    raise ValueError('Namespace %s not available' % namespace)
ValueError: Namespace Polkit not available

Fixes #613