redefine --followlinks
Make --followlinks treat symlinks to files as copies of said files
instead of ways to traverse files outside of searched paths. In other
words, it references the path of symlinks instead of their targets in
the output.

Also, add a --keep-symlinks option and make it the default when
--followlinks is given.

Fixes #586
Fixes #589