replay: fix incorrect use of S_IFLNK
S_IFLNK is not a mask representing a bit in a bitfield. Rather, it is
the value of `st_mode & S_IFMT` for symbolic links. This was causing
regular files to be detected as symbolic links, though this was harmless
because replay currently doesn't do anything with that information.

Use the S_ISLNK macro instead.