utils: compare logical extent offset in rm_util_link_type
Holes are handled by ignoring them when they are after an extent and
skipping them when they are before one. Since we were only comparing the
current file position, a file with a hole could match a file without one
as long as the physical offsets match and the extents logically end at
the same place. This led to --is-reflink false-positives.

Fix this by comparing the logical offset of each extent.

Fixes #611