compat: simplify LFS
Remove references to stat64.

Restore _FILE_OFFSET_BITS=64 everywhere. It is really only needed
for glibc 32-bit as the other supported platforms use 'stat'
with a 64-bit API.

_FILE_OFFSET_BITS=64 was dropped in ebbec1f (year 2014), citing an
incompatibility with libelf. However, in 2015 elfutils 0.164 moved
the ABI from (l)off_t to int64_t. That, and the fact that we never
used libelf functions that used offsets, and the fact that FreeBSD
has always defined off_t as 64-bit (since 1991) leads to the
conclusion that we can revert the 12 years old decision that
needed many workarounds.

Added a check to ensure a compatible libelf.